refactor: combine unstructured and resource packages
