+++ cmd/cli/kubectl-kyverno/commands/test/compare.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/go-git/go-billy/v5"
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/resource"
-	unstructuredutils "github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/unstructured"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
@@ -14,9 +13,9 @@ func getAndCompareResource(actualResource unstructured.Unstructured, fs billy.Fi
 	if err != nil {
 		return false, fmt.Errorf("Error: failed to load resource (%s)", err)
 	}
-	unstructuredutils.FixupGenerateLabels(actualResource)
-	unstructuredutils.FixupGenerateLabels(*expectedResource)
-	equals, err := unstructuredutils.Compare(actualResource, *expectedResource, true)
+	resource.FixupGenerateLabels(actualResource)
+	resource.FixupGenerateLabels(*expectedResource)
+	equals, err := resource.Compare(actualResource, *expectedResource, true)
 	if err != nil {
 		return false, fmt.Errorf("Error: failed to compare resources (%s)", err)
 	}
